qtl_experiments/requirements.txt
torch>=2.0
torchvision>=0.15
pennylane>=0.38
pennylane-lightning>=0.38
qiskit>=1.3
qiskit-aer>=0.17
qiskit-machine-learning>=0.8
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
codecarbon>=2.3
tqdm>=4.65
pillow>=9.0
